Elimina variable DataPuntosEncuentro Limpia el código

+++ app/Http/Controllers/backoffice/ActividadesController.php
@@ -115,7 +115,9 @@ class ActividadesController extends Controller
                 'tipo.categoria',
                 'oficina',
                 'modificadoPor',
-                'puntosEncuentro',
+                'puntosEncuentro.pais',
+                'puntosEncuentro.provincia',
+                'puntosEncuentro.localidad',
                 'pais',
                 'provincia',
                 'localidad',
@@ -378,7 +380,8 @@ class ActividadesController extends Controller
             if (!empty($request->puntos_encuentro)) {
                 $puntosGuardados = $actividad->puntosEncuentro->count() > 0 ? $actividad->puntosEncuentro->pluck('idPuntoEncuentro')->toArray() : [];
                 foreach ($request->puntos_encuentro as $punto) {
-                    if (!in_array($punto['idPuntoEncuentro'], $puntosGuardados)) {
+//                    if (!in_array($punto['idPuntoEncuentro'], $puntosGuardados)) {
+                    if (!empty($punto['nuevo']) && $punto['nuevo'] == true) {
                         $punto = $this->guardarPunto($punto, $actividad);
                     }
                 }
@@ -386,7 +389,9 @@ class ActividadesController extends Controller
 
             foreach ($request->puntosEncuentroBorrados as $borrado) {
                 $punto = PuntoEncuentro::find($borrado['idPuntoEncuentro']);
-                $punto->delete();
+                if ($punto) {
+                    $punto->delete();
+                }
             }
             return true;
         }
